fix(observability): address PR review comments on tracing instrumentation
- Add explicit base_branch and base_sha to the instrument fields(...) list
  in ReleaseOrchestrator::orchestrate for clarity and intent alignment
- Add #[instrument(skip(self))] to seven GitHubClient methods that were
  emitting structured log events but lacking child spans:
  create_branch, delete_branch, create_issue_comment,
  get_collaborator_permission, add_labels, remove_label, list_pr_labels
- Split structured-logging span tests out of release_orchestrator_tests.rs
  into release_orchestrator_tracing_tests.rs to keep the main file under
  the 500-line policy limit (.tech-decisions.yml)
