+++ crates/core/src/release_orchestrator_tests.rs
@@ -16,7 +16,6 @@ use async_trait::async_trait;
 use chrono::Utc;
 use std::sync::Arc;
 use tokio::sync::Mutex;
-use tracing_test::traced_test;
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Inline test double
@@ -49,12 +48,12 @@ struct TestState {
 }
 
 #[derive(Clone, Default)]
-struct TestGitHub {
+pub(super) struct TestGitHub {
     state: Arc<Mutex<TestState>>,
 }
 
 impl TestGitHub {
-    fn new() -> Self {
+    pub(super) fn new() -> Self {
         Self {
             state: Arc::new(Mutex::new(TestState {
                 next_pr_number: 100,
@@ -403,7 +402,7 @@ impl GitHubOperations for TestGitHub {
 }
 // ─────────────────────────────────────────────────────────────────────────────
 
-fn stub_repo(owner: &str, repo: &str) -> Repository {
+pub(super) fn stub_repo(owner: &str, repo: &str) -> Repository {
     Repository {
         id: 1,
         name: repo.to_string(),
@@ -418,7 +417,7 @@ fn stub_repo(owner: &str, repo: &str) -> Repository {
     }
 }
 
-fn stub_git_user() -> GitHubUser {
+pub(super) fn stub_git_user() -> GitHubUser {
     GitHubUser {
         name: "test-user".to_string(),
         email: "test@example.com".to_string(),
@@ -426,7 +425,7 @@ fn stub_git_user() -> GitHubUser {
     }
 }
 
-fn make_open_release_pr(number: u64, branch: &str, body: Option<&str>) -> PullRequest {
+pub(super) fn make_open_release_pr(number: u64, branch: &str, body: Option<&str>) -> PullRequest {
     let now = Utc::now();
     let r = stub_repo("testorg", "testrepo");
     PullRequest {
@@ -452,7 +451,7 @@ fn make_open_release_pr(number: u64, branch: &str, body: Option<&str>) -> PullRe
     }
 }
 
-fn ver(major: u64, minor: u64, patch: u64) -> SemanticVersion {
+pub(super) fn ver(major: u64, minor: u64, patch: u64) -> SemanticVersion {
     SemanticVersion {
         major,
         minor,
@@ -462,9 +461,17 @@ fn ver(major: u64, minor: u64, patch: u64) -> SemanticVersion {
     }
 }
 
-fn default_config() -> OrchestratorConfig {
+pub(super) fn default_config() -> OrchestratorConfig {
     OrchestratorConfig::default()
 }
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Structured logging sub-module
+// ─────────────────────────────────────────────────────────────────────────────
+
+#[cfg(test)]
+#[path = "release_orchestrator_tracing_tests.rs"]
+mod tracing_tests;
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Tests
@@ -1031,96 +1038,4 @@ fn test_extract_changelog_from_pr_body_empty_section_returns_empty() {
 fn test_extract_changelog_from_pr_body_empty_body_returns_empty() {
     let result = extract_changelog_from_pr_body("", "## Changelog");
     assert_eq!(result, "");
-}
-
-// ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-// Structured logging / span propagation tests
-// ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-/// `orchestrate` must include the `correlation_id` as a structured field in its
-/// span so that ALL log lines emitted during orchestration are correlated.
-///
-/// This test will fail if the function uses `span.enter()` without also
-/// including `correlation_id` in a log event, or if `#[tracing::instrument]`
-/// is absent and the span is never opened.
-#[tokio::test]
-#[traced_test]
-async fn test_orchestrate_logs_contain_correlation_id() {
-    let github = TestGitHub::new();
-    let orchestrator = ReleaseOrchestrator::new(default_config(), &github);
-
-    orchestrator
-        .orchestrate(
-            "testorg",
-            "testrepo",
-            &ver(1, 0, 0),
-            "- feat: something",
-            "main",
-            "sha-001",
-            "unique-corr-id-ALPHA",
-        )
-        .await
-        .expect("orchestrate should succeed");
-
-    assert!(
-        logs_contain("unique-corr-id-ALPHA"),
-        "correlation_id must appear in structured log output"
-    );
-}
-
-/// `orchestrate` must record `owner` and `repo` as structured fields so that
-/// log lines are filterable by repository in production monitoring.
-#[tokio::test]
-#[traced_test]
-async fn test_orchestrate_logs_contain_owner_and_repo() {
-    let github = TestGitHub::new();
-    let orchestrator = ReleaseOrchestrator::new(default_config(), &github);
-
-    orchestrator
-        .orchestrate(
-            "unique-owner-BETA",
-            "unique-repo-BETA",
-            &ver(2, 0, 0),
-            "- fix: something",
-            "main",
-            "sha-002",
-            "corr-002",
-        )
-        .await
-        .expect("orchestrate should succeed");
-
-    assert!(
-        logs_contain("unique-owner-BETA"),
-        "owner must appear in structured log output"
-    );
-    assert!(
-        logs_contain("unique-repo-BETA"),
-        "repo must appear in structured log output"
-    );
-}
-
-/// `orchestrate` must record the version as a structured field.
-#[tokio::test]
-#[traced_test]
-async fn test_orchestrate_logs_contain_version() {
-    let github = TestGitHub::new();
-    let orchestrator = ReleaseOrchestrator::new(default_config(), &github);
-
-    orchestrator
-        .orchestrate(
-            "testorg",
-            "testrepo",
-            &ver(9, 8, 7),
-            "- chore: bump",
-            "main",
-            "sha-003",
-            "corr-003",
-        )
-        .await
-        .expect("orchestrate should succeed");
-
-    assert!(
-        logs_contain("9.8.7"),
-        "version must appear in structured log output"
-    );
 }
